# Backend/SwaggerDocs/api.yaml
openapi: 3.0.1
info:
  title: Appointments API
  description: API for managing appointments
  version: v1
paths:
  /api/Appointments:
    get:
      tags:
        - Appointments
      parameters:
        - name: date
          in: query
          schema:
            type: string
            format: date-time
      responses:
        '200':
          description: Success
    post:
      tags:
        - Appointments
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Appointment'
          text/json:
            schema:
              $ref: '#/components/schemas/Appointment'
          application/*+json:
            schema:
              $ref: '#/components/schemas/Appointment'
      responses:
        '200':
          description: Success
  '/api/Appointments/{id}':
    put:
      tags:
        - Appointments
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int32
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Appointment'
          text/json:
            schema:
              $ref: '#/components/schemas/Appointment'
          application/*+json:
            schema:
              $ref: '#/components/schemas/Appointment'
      responses:
        '200':
          description: Success
    delete:
      tags:
        - Appointments
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int32
      responses:
        '200':
          description: Success
  /api/Appointments/upcoming:
    get:
      tags:
        - Appointments
      parameters:
        - name: start
          in: query
          schema:
            type: string
            format: date-time
        - name: end
          in: query
          schema:
            type: string
            format: date-time
      responses:
        '200':
          description: Success
components:
  schemas:
    Appointment:
      required:
        - endTime
        - startTime
      type: object
      properties:
        id:
          type: integer
          format: int32
        title:
          maxLength: 100
          minLength: 0
          type: string
          nullable: true
        description:
          maxLength: 500
          minLength: 0
          type: string
          nullable: true
        startTime:
          type: string
          format: date-time
        endTime:
          type: string
          format: date-time
        createdAt:
          type: string
          format: date-time
      additionalProperties: false